Fixed how PRIV tags are handled.
Some software (i.e. Windows) adds funky tags with binary data in the
form of PRIV tags.  Those tags do NOT render well.  I've changed it so
the tag owner field is printed, but not the data itself.

Some time soon I'll add a way to force it to print it all anyway.